requirements.txt
numpy == 1.26.0
matplotlib == 3.8.0
pyyaml == 6.0.1